[PERF-862] captcha from v0.0.3 (#2)
* Add a block around the contact submit action

Allows extension of this block by other extensions.

* Reformat html file

* Add block around recaptcha

* More reformatting

* Reformat contact controller a bit

* Add requests as a dependency

* Some light formatting changes

* Add recaptcha support to contact forms

* Refactor to allow ajax form submission to work correctly

This commit just contains the necessary changes, there will be a
refactoring commit afterwards to make some non-function changing changes
in the interest of the code being nice to work with.

* General tidy up, unicode and doc

* Initialise the recaptcha context after the modal is launched

By doing this when the modal is shown we avoid showing the recaptcha
badge on the page before the user has even given an indication that they
want to contact us which avoids confusion.

* Tidy up the modal javascript module

* More minor doc edits in the javascript

* Small refactor to check the recaptcha key and action correctly

Prior to this when no key or action was available the code didn't handle
it correctly. Now it does. The refactored check and context creation
also centralises the logic so that the recaptcha helpers handle the
checking.

* Renable sending emails

* Update README.md

* Fix merge result

* Rename contact field names

Co-authored-by: Josh Humphries <j.humphries@nhm.ac.uk>
Co-authored-by: Josh Humphries <jrdhumphries@gmail.com>
Co-authored-by: Jay Guo <jguo144@gmail.com>